+++ src/main/java/com/sparta/haengye_project/user/security/UserDetailsImpl.java
@@ -1,4 +1,4 @@
-package com.sparta.haengye_project.security;
+package com.sparta.haengye_project.user.security;
 
 import com.sparta.haengye_project.user.entity.User;
 import org.springframework.security.core.GrantedAuthority;
