+++ src/main/java/com/sparta/haengye_project/user/security/UserDetailsServiceImpl.java
@@ -1,4 +1,4 @@
-package com.sparta.haengye_project.security;
+package com.sparta.haengye_project.user.security;
 
 import com.sparta.haengye_project.user.entity.User;
 import com.sparta.haengye_project.user.repository.UserRepository;
